Copies emoji on click.
The event listener is placed on the `mat-grid-list` element rather than the `mat-grid-tile` to use only one event listener instead of O(n) listeners. This gets a little awkward because does not retarget events, so the event target is actually a `<div />` under `mat-grid-tile`. I had to manually walk the DOM tree up until a found the right element to get the text from.