Adds `Noto Color Emoji` font to the page for browsers which don't have it installed.
Windows Chrome uses Segoe Emoji UI, which as of now is missing many glyphs and renders several glyphs as a combination of two. For instance, the "heart on fire" emoji is displayed like a "heart" emoji followed by a "fire" emoji. This means the width of emojis is highly variable and difficult to display in an aesthetically pleasing fashion.

Font downloaded from: https://github.com/googlefonts/noto-emoji.